admin_panel/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
PyJWT==2.3.0
Flask-JWT-Extended==4.4.4
Pillow==9.0.0
mutagen==1.45.1
Werkzeug==2.0.1
python-dotenv==0.19.0 